Reorganize directory structure to work with recent vim versions
This reorgnization allows for one to clone the repository directly to
`~/.vim` without needing any extra symlinks from `.vim_dotfiles` or
other location into one's home directory. All that's needed from then is
to run `vim +PlugInstall +PlugClean! +qall` and you have a fully
functioning vim setup, self contained in `~/.vim`.

The `activate.sh` script left for consistency's sake and to not break
people's workflows. It may also want to warn about legacy symlinks.

+++ activate.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 
 dotfiles_dir=$(cd "$(dirname "$0")"; pwd)
 
-for name in vim vimrc vimrc.bundles; do
-  rm -rf "${HOME}/.${name}"
-  ln -s "${dotfiles_dir}/${name}" "${HOME}/.${name}"
-done
+if ! vim --version | grep -q '2nd user vimrc file'; then
+  for name in vim vimrc vimrc.bundles; do
+    rm -rf "${HOME}/.${name}"
+    ln -s "${dotfiles_dir}/${name}" "${HOME}/.${name}"
+  done
+fi
 
 vim +PlugInstall +PlugClean! +qall
